[ALPS07779043] vcodec: fix build error
Type error introduced by new clang

MTK-Commit-Id: 2b90c134056332a3881f9329db02cb506bd8b157

CR-Id: ALPS07779043
Feature: Kernel SI Operation

+++ drivers/media/platform/mtk-vcodec/v1/mtk_vcodec_enc_pm_plat.c
@@ -382,14 +382,14 @@ void set_venc_opp(struct mtk_vcodec_dev *dev, u32 freq)
 				mtk_v4l2_err("[VENC] Failed to set mmdvfs rate %lu\n",
 						freq_64);
 			}
-			mtk_v4l2_debug(4, "[VENC] freq %lu, find_freq %lu", freq, freq_64);
+			mtk_v4l2_debug(4, "[VENC] freq %u, find_freq %lu", freq, freq_64);
 		} else if (dev->venc_reg) {
 			ret = regulator_set_voltage(dev->venc_reg, volt, INT_MAX);
 			if (ret) {
 				mtk_v4l2_err("[VENC] Failed to set regulator voltage %d\n",
 						volt);
 			}
-			mtk_v4l2_debug(4, "[VENC] freq %lu, voltage %lu", freq, volt);
+			mtk_v4l2_debug(4, "[VENC] freq %u, voltage %d", freq, volt);
 		}
 	}
 }
@@ -432,7 +432,7 @@ void mtk_venc_pmqos_begin_inst(struct mtk_vcodec_ctx *ctx)
 		if (dev->venc_port_bw[i].port_type < VCODEC_PORT_LARB_SUM) {
 			mtk_icc_set_bw_not_update(dev->venc_qos_req[i],
 					MBps_to_icc((u32)target_bw), 0);
-			mtk_v4l2_debug(8, "[VENC] port %d bw %lu MB/s", i, (u32)target_bw);
+			mtk_v4l2_debug(8, "[VENC] port %d bw %llu MB/s", i, target_bw);
 		} else if (dev->venc_port_bw[i].port_type == VCODEC_PORT_LARB_SUM) {
 			mtk_icc_set_bw(dev->venc_qos_req[i], 0, 0);
 			mtk_v4l2_debug(8, "[VENC] port %d set larb %u bw",
@@ -463,7 +463,7 @@ void mtk_venc_pmqos_end_inst(struct mtk_vcodec_ctx *ctx)
 		if (dev->venc_port_bw[i].port_type < VCODEC_PORT_LARB_SUM) {
 			mtk_icc_set_bw_not_update(dev->venc_qos_req[i],
 					MBps_to_icc((u32)target_bw), 0);
-			mtk_v4l2_debug(8, "[VENC] port %d bw %lu MB/s", i, (u32)target_bw);
+			mtk_v4l2_debug(8, "[VENC] port %d bw %llu MB/s", i, target_bw);
 		} else if (dev->venc_port_bw[i].port_type == VCODEC_PORT_LARB_SUM) {
 			mtk_icc_set_bw(dev->venc_qos_req[i], 0, 0);
 			mtk_v4l2_debug(8, "[VENC] port %d set larb %u bw",
